backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pydantic>=2.8.0
email-validator>=2.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
jinja2==3.1.2
jsonschema==4.20.0
mcp==1.0.0
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
weasyprint==62.3
pydyf==0.10.0
